yolo_clip_detector/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.18.0
opencv-python>=4.5.1
Pillow>=8.0.0
PyYAML>=5.4.1
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
transformers>=4.11.0
timm>=0.5.4
ftfy>=6.0.3
regex
tensorboard>=2.5.0
pycocotools>=2.0.2
onnx>=1.10.0
onnxruntime>=1.8.0